KRACOEUS-6955 : common rename question to questionnaire
