+++ coeus-code/src/main/java/org/kuali/coeus/common/committee/impl/web/struts/action/CommitteeActionsActionBase.java
@@ -38,7 +38,7 @@ import org.kuali.coeus.common.committee.impl.web.struts.form.CommitteeFormBase;
 import org.kuali.coeus.common.framework.print.AbstractPrint;
 import org.kuali.coeus.common.framework.print.Printable;
 import org.kuali.coeus.common.framework.print.PrintingException;
-import org.kuali.coeus.common.question.framework.print.CorrespondencePrintingService;
+import org.kuali.coeus.common.questionnaire.framework.print.CorrespondencePrintingService;
 import org.kuali.kra.infrastructure.Constants;
 import org.kuali.kra.infrastructure.TaskName;
 import org.kuali.coeus.common.framework.print.AttachmentDataSource;
