+++ coeus-code/src/main/java/org/kuali/coeus/common/questionnaire/framework/print/CorrespondencePrint.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.coeus.common.question.framework.print;
+package org.kuali.coeus.common.questionnaire.framework.print;
 
 import java.io.ByteArrayInputStream;
 import java.util.ArrayList;
