+++ coeus-code/src/main/java/org/kuali/coeus/common/questionnaire/framework/print/QuestionnairePrintingService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.coeus.common.question.framework.print;
+package org.kuali.coeus.common.questionnaire.framework.print;
 
 import org.kuali.coeus.common.framework.print.Printable;
 import org.kuali.coeus.common.framework.print.PrintingException;
