+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/print/ProposalDevelopmentPrintingServiceImpl.java
@@ -32,7 +32,7 @@ import org.kuali.coeus.propdev.impl.core.DevelopmentProposal;
 import org.kuali.coeus.propdev.impl.person.ProposalPerson;
 import org.kuali.coeus.propdev.impl.person.question.ProposalPersonQuestionnaireHelper;
 import org.kuali.kra.questionnaire.answer.AnswerHeader;
-import org.kuali.coeus.common.question.framework.print.QuestionnairePrint;
+import org.kuali.coeus.common.questionnaire.framework.print.QuestionnairePrint;
 import org.kuali.rice.coreservice.framework.parameter.ParameterService;
 import org.kuali.rice.krad.service.BusinessObjectService;
