+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/question/ProposalDevelopmentQuestionsAction.java
@@ -28,7 +28,7 @@ import org.kuali.coeus.common.framework.print.AttachmentDataSource;
 import org.kuali.kra.questionnaire.answer.AnswerHeader;
 import org.kuali.kra.questionnaire.answer.QuestionnaireAnswerService;
 import org.kuali.kra.questionnaire.answer.SaveQuestionnaireAnswerEvent;
-import org.kuali.coeus.common.question.framework.print.QuestionnairePrintingService;
+import org.kuali.coeus.common.questionnaire.framework.print.QuestionnairePrintingService;
 import org.kuali.rice.krad.document.Document;
 
 import javax.servlet.http.HttpServletRequest;
