+++ coeus-code/src/main/java/org/kuali/kra/committee/service/impl/ScheduleCorrespondencePrintImpl.java
@@ -16,7 +16,7 @@
 package org.kuali.kra.committee.service.impl;
 
 import org.kuali.kra.committee.service.CommitteeCorrespondencePrint;
-import org.kuali.coeus.common.question.framework.print.CorrespondencePrintingServiceImpl;
+import org.kuali.coeus.common.questionnaire.framework.print.CorrespondencePrintingServiceImpl;
 import org.kuali.kra.iacuc.IacucProtocol;
 import org.kuali.kra.iacuc.actions.submit.IacucProtocolSubmission;
 import org.kuali.kra.protocol.ProtocolBase;
