+++ coeus-code/src/main/java/org/kuali/kra/committee/web/struts/action/CommitteeActionsAction.java
@@ -28,7 +28,7 @@ import org.kuali.kra.committee.rule.event.CommitteeActionGenerateBatchCorrespond
 import org.kuali.kra.committee.service.CommitteeBatchCorrespondenceService;
 import org.kuali.kra.committee.service.CommitteeCorrespondencePrint;
 import org.kuali.kra.committee.service.CommitteePrintingService;
-import org.kuali.coeus.common.question.framework.print.CorrespondencePrintingService;
+import org.kuali.coeus.common.questionnaire.framework.print.CorrespondencePrintingService;
 import org.kuali.kra.infrastructure.TaskGroupName;
 
 import java.sql.Date;
