+++ coeus-code/src/main/java/org/kuali/kra/iacuc/questionnaire/print/IacucCorrespondencePrintingService.java
@@ -15,7 +15,7 @@
  */
 package org.kuali.kra.iacuc.questionnaire.print;
 
-import org.kuali.coeus.common.question.framework.print.CorrespondencePrintingService;
+import org.kuali.coeus.common.questionnaire.framework.print.CorrespondencePrintingService;
 
 public interface IacucCorrespondencePrintingService extends CorrespondencePrintingService {
