+++ coeus-code/src/main/java/org/kuali/kra/iacuc/questionnaire/print/QuestionnairePrintingServiceImpl.java
@@ -24,7 +24,7 @@ import org.kuali.kra.protocol.ProtocolBase;
 import org.kuali.kra.protocol.actions.print.QuestionnairePrintOption;
 import org.kuali.kra.protocol.actions.submit.ProtocolSubmissionBase;
 import org.kuali.kra.questionnaire.Questionnaire;
-import org.kuali.coeus.common.question.framework.print.QuestionnairePrint;
+import org.kuali.coeus.common.questionnaire.framework.print.QuestionnairePrint;
 import org.kuali.rice.krad.service.BusinessObjectService;
 
 import java.util.ArrayList;
