+++ coeus-code/src/main/java/org/kuali/kra/irb/ProtocolAction.java
@@ -49,7 +49,7 @@ import org.kuali.kra.protocol.auth.ProtocolTaskBase;
 import org.kuali.kra.protocol.notification.ProtocolNotification;
 import org.kuali.kra.protocol.notification.ProtocolNotificationContextBase;
 import org.kuali.kra.questionnaire.answer.AnswerHeader;
-import org.kuali.coeus.common.question.framework.print.QuestionnairePrintingService;
+import org.kuali.coeus.common.questionnaire.framework.print.QuestionnairePrintingService;
 import org.kuali.rice.kew.api.KewApiConstants;
 import org.kuali.rice.krad.document.Document;
 import org.kuali.rice.krad.service.KRADServiceLocatorWeb;
