+++ coeus-code/src/main/java/org/kuali/kra/irb/questionnaire/print/IrbCorrespondencePrintingService.java
@@ -15,7 +15,7 @@
  */
 package org.kuali.kra.irb.questionnaire.print;
 
-import org.kuali.coeus.common.question.framework.print.CorrespondencePrintingService;
+import org.kuali.coeus.common.questionnaire.framework.print.CorrespondencePrintingService;
 
 public interface IrbCorrespondencePrintingService extends CorrespondencePrintingService{
