+++ coeus-code/src/main/java/org/kuali/kra/questionnaire/QuestionnaireMaintenanceDocumentAction.java
@@ -29,7 +29,7 @@ import org.kuali.kra.infrastructure.Constants;
 import org.kuali.kra.infrastructure.KeyConstants;
 import org.kuali.kra.infrastructure.PermissionConstants;
 import org.kuali.coeus.common.framework.print.AttachmentDataSource;
-import org.kuali.coeus.common.question.framework.print.QuestionnairePrintingService;
+import org.kuali.coeus.common.questionnaire.framework.print.QuestionnairePrintingService;
 import org.kuali.kra.questionnaire.question.Question;
 import org.kuali.rice.kns.document.MaintenanceDocumentBase;
 import org.kuali.rice.kns.question.ConfirmationQuestion;
