+++ coeus-it/src/test/java/org/kuali/kra/questionnaire/print/QuestionnairePrintingServiceTest.java
@@ -22,9 +22,9 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.kuali.coeus.common.framework.print.AbstractPrint;
 import org.kuali.coeus.common.framework.print.Printable;
-import org.kuali.coeus.common.question.framework.print.QuestionnairePrint;
-import org.kuali.coeus.common.question.framework.print.QuestionnairePrintingService;
-import org.kuali.coeus.common.question.impl.print.QuestionnairePrintingServiceImpl;
+import org.kuali.coeus.common.questionnaire.framework.print.QuestionnairePrint;
+import org.kuali.coeus.common.questionnaire.framework.print.QuestionnairePrintingService;
+import org.kuali.coeus.common.questionnaire.impl.print.QuestionnairePrintingServiceImpl;
 import org.kuali.coeus.sys.framework.service.KcServiceLocator;
 import org.kuali.kra.irb.Protocol;
 import org.kuali.kra.irb.ProtocolDocument;
